optimize the vector
The vector implementation was using initialized memory to store objects
thus calling the default constructor *capacity* times every time it was
constructed or resized. This could have tanked it's performance if it
was used to store large or expensively constructed T's (not even
mentioning that it wouldn't work with classes that didn't have a default
constructor). This optimization uses std::aligned_storage to create a
data type with the correct size and alignment for uninitialized memory
allocation thus avoiding all the pitfalls mentioned above